Move GQL functional tests to their own suite.
