+++ tests/_support/Helper/Gql.php
@@ -1,0 +1,20 @@
+<?php
+namespace Helper;
+
+use Codeception\Module;
+
+/**
+ * Class Gql
+ *
+ * Here you can define custom actions.
+ * All public methods declared in helper class will be available in $I
+ *
+ * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
+ * @author Global Network Group | Giel Tettelaar <giel@yellowflash.net>
+ * @since 3.3
+ */
+
+class Gql extends Module
+{
+
+}
