+++ tests/gql/_bootstrap.php
@@ -1,0 +1,5 @@
+<?php
+// Autoload the fixtures
+use Codeception\Util\Autoload;
+
+Autoload::addNamespace('', dirname(__DIR__) . DIRECTORY_SEPARATOR . 'fixtures');
